chore: update exported types
