+++ src/index.ts
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
+import path from 'path';
+import os from 'os';
+import fs from 'fs';
+
 import { createServerWithTools } from './server';
 import common from './tools/common';
 import files from './tools/files';
@@ -58,9 +62,10 @@ const resources: Resource[] = [
 ];
 
 type Options = {
-  browserName?: 'chromium' | 'firefox' | 'webkit';
+  browser?: string;
   userDataDir?: string;
-  launchOptions?: LaunchOptions;
+  headless?: boolean;
+  executablePath?: string;
   cdpEndpoint?: string;
   vision?: boolean;
   capabilities?: ToolCapability[];
@@ -68,7 +73,42 @@ type Options = {
 
 const packageJSON = require('../package.json');
 
-export function createServer(options?: Options): Server {
+export async function createServer(options?: Options): Promise<Server> {
+  let browserName: 'chromium' | 'firefox' | 'webkit';
+  let channel: string | undefined;
+  switch (options?.browser) {
+    case 'chrome':
+    case 'chrome-beta':
+    case 'chrome-canary':
+    case 'chrome-dev':
+    case 'msedge':
+    case 'msedge-beta':
+    case 'msedge-canary':
+    case 'msedge-dev':
+      browserName = 'chromium';
+      channel = options.browser;
+      break;
+    case 'chromium':
+      browserName = 'chromium';
+      break;
+    case 'firefox':
+      browserName = 'firefox';
+      break;
+    case 'webkit':
+      browserName = 'webkit';
+      break;
+    default:
+      browserName = 'chromium';
+      channel = 'chrome';
+  }
+  const userDataDir = options?.userDataDir ?? await createUserDataDir(browserName);
+
+  const launchOptions: LaunchOptions = {
+    headless: !!(options?.headless ?? (os.platform() === 'linux' && !process.env.DISPLAY)),
+    channel,
+    executablePath: options?.executablePath,
+  };
+
   const allTools = options?.vision ? screenshotTools : snapshotTools;
   const tools = allTools.filter(tool => !options?.capabilities || tool.capability === 'core' || options.capabilities.includes(tool.capability));
   return createServerWithTools({
@@ -76,9 +116,24 @@ export function createServer(options?: Options): Server {
     version: packageJSON.version,
     tools,
     resources,
-    browserName: options?.browserName,
-    userDataDir: options?.userDataDir ?? '',
-    launchOptions: options?.launchOptions,
+    browserName,
+    userDataDir,
+    launchOptions,
     cdpEndpoint: options?.cdpEndpoint,
   });
+}
+
+async function createUserDataDir(browserName: 'chromium' | 'firefox' | 'webkit') {
+  let cacheDirectory: string;
+  if (process.platform === 'linux')
+    cacheDirectory = process.env.XDG_CACHE_HOME || path.join(os.homedir(), '.cache');
+  else if (process.platform === 'darwin')
+    cacheDirectory = path.join(os.homedir(), 'Library', 'Caches');
+  else if (process.platform === 'win32')
+    cacheDirectory = process.env.LOCALAPPDATA || path.join(os.homedir(), 'AppData', 'Local');
+  else
+    throw new Error('Unsupported platform: ' + process.platform);
+  const result = path.join(cacheDirectory, 'ms-playwright', `mcp-${browserName}-profile`);
+  await fs.promises.mkdir(result, { recursive: true });
+  return result;
 }
